+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog for tasty-quickcheck-laws
 ===================================
 
+Unreleased changes
+------------------
+
+* Fixed
+    * Clarified license (it's BSD3)
+
 0.0.1
 -----